# docs/openapi.yml
{
  "openapi": "3.0.2",
  "info": {
    "title": "Observer API",
    "description": "Observer API server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/support-records": {
      "post": {
        "tags": [
          "support records"
        ],
        "summary": "Create Support Record",
        "operationId": "create_support_record_support_records_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewSupportRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportRecordResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/support-records/{record_id}": {
      "get": {
        "tags": [
          "support records"
        ],
        "summary": "Get Support Record",
        "operationId": "get_support_record_support_records__record_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Record Id",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "record_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportRecordResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "support records"
        ],
        "summary": "Update Support Record",
        "operationId": "update_support_record_support_records__record_id__put",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Record Id",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "record_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupportRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportRecordResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "support records"
        ],
        "summary": "Delete Support Record",
        "operationId": "delete_support_record_support_records__record_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Record Id",
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "name": "record_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "title": "Access Token",
              "type": "string"
            },
            "name": "access_token",
            "in": "cookie"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "title": "APIError",
        "required": [
          "code",
          "status_code",
          "message"
        ],
        "type": "object",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "status_code": {
            "title": "Status Code",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "data": {
            "title": "Data"
          }
        }
      },
      "AgeGroup": {
        "title": "AgeGroup",
        "enum": [
          "infant",
          "toddler",
          "pre_school",
          "middle_childhood",
          "young_teen",
          "teenager",
          "young_adult",
          "early_adult",
          "middle_aged_adult",
          "old_adult",
          "unknown"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
      "ErrorCode": {
        "title": "ErrorCode",
        "enum": [
          "unauthorized",
          "forbidden",
          "not_found",
          "conflict_error",
          "totp_error",
          "totp_invalid_backup_code_error",
          "totp_required_error",
          "totp_exists_error",
          "registration_error",
          "registrations_closed_error",
          "password_reset_code_expired_error",
          "weak_password_error",
          "invalid_password_error",
          "similar_passwords_error",
          "document_is_too_large_error",
          "unsupported_document_format",
          "confirmation_code_expired_error",
          "invite_expired_error",
          "internal_error",
          "bad_request"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "NewSupportRecordRequest": {
        "title": "NewSupportRecordRequest",
        "required": [
          "type",
          "consultant_id",
          "record_for",
          "owner_id",
          "project_id"
        ],
        "type": "object",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string",
            "description": "Description of support record"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SupportType"
              }
            ],
            "description": "Type of support"
          },
          "consultant_id": {
            "title": "Consultant Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Consultant ID"
          },
          "age_group": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AgeGroup"
              }
            ],
            "description": "Beneficiary age group"
          },
          "record_for": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SupportRecordSubject"
              }
            ],
            "description": "Record subject"
          },
          "owner_id": {
            "title": "Owner Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Owner of support humans or pets"
          },
          "project_id": {
            "title": "Project Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Project ID"
          }
        }
      },
...
      "SupportRecordResponse": {
        "title": "SupportRecordResponse",
        "required": [
          "type",
          "consultant_id",
          "record_for",
          "owner_id",
          "project_id",
          "id",
          "created_at"
        ],
        "type": "object",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string",
            "description": "Description of support record"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SupportType"
              }
            ],
            "description": "Type of support"
          },
          "consultant_id": {
            "title": "Consultant Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Consultant ID"
          },
          "age_group": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AgeGroup"
              }
            ],
            "description": "Beneficiary age group"
          },
          "record_for": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SupportRecordSubject"
              }
            ],
            "description": "Record subject"
          },
          "owner_id": {
            "title": "Owner Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Owner of support humans or pets"
          },
          "project_id": {
            "title": "Project Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Project ID"
          },
          "id": {
            "title": "Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Support record ID"
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "description": "Creation date",
            "format": "date-time"
          }
        }
      },
      "SupportRecordSubject": {
        "title": "SupportRecordSubject",
        "enum": [
          "person",
          "pet"
        ],
        "type": "string",
        "description": "An enumeration."
      },
      "SupportType": {
        "title": "SupportType",
        "enum": [
          "humanitarian",
          "legal",
          "medical",
          "general"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
      "UpdateSupportRecordRequest": {
        "title": "UpdateSupportRecordRequest",
        "required": [
          "type",
          "consultant_id",
          "record_for",
          "owner_id",
          "project_id"
        ],
        "type": "object",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string",
            "description": "Description of support record"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SupportType"
              }
            ],
            "description": "Type of support"
          },
          "consultant_id": {
            "title": "Consultant Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Consultant ID"
          },
          "age_group": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AgeGroup"
              }
            ],
            "description": "Beneficiary age group"
          },
          "record_for": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SupportRecordSubject"
              }
            ],
            "description": "Record subject"
          },
          "owner_id": {
            "title": "Owner Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Owner of support humans or pets"
          },
          "project_id": {
            "title": "Project Id",
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "description": "Project ID"
          }
        }
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}